Refactor data helper function
- Rename src/helpers/data.js to src/helpers/state.js
- Update paths to import statements
