+++ src/HOCs/routes/AuthRoute.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect, Route } from 'react-router-dom';
 
-import { generateState } from '../../helpers/data';
+import { generateState } from '../../helpers/state';
 import toJS from '../../HOCs/state/toJS';
 import { getAuth } from '../../data/session/reducers';
