+++ src/screens/Surveys/screens/SurveyDetails/data/survey/sagas.js
@@ -4,7 +4,7 @@ import { mapKeys } from 'lodash';
 import { all, call, put, takeLatest } from 'redux-saga/effects';
 
 // Helper functions and services
-import { fromJSOrdered, getError } from '../../../../../../helpers/data';
+import { fromJSOrdered, getError } from '../../../../../../helpers/state';
 import * as surveysService from '../../../../../../services/surveys';
 
 // Action types and action creators
