+++ src/screens/Surveys/screens/SurveyDoorway/data/landing/sagas.js
@@ -3,7 +3,7 @@ import { fromJS } from 'immutable';
 import { call, put, takeLatest } from 'redux-saga/effects';
 
 // Helper functions and services
-import { getError } from '../../../../../../helpers/data';
+import { getError } from '../../../../../../helpers/state';
 import * as surveysService from '../../../../../../services/surveys';
 
 // Action types and action creators
