+++ src/screens/Surveys/screens/SurveyDoorway/index.js
@@ -2,7 +2,7 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { generateState } from '../../../../helpers/data';
+import { generateState } from '../../../../helpers/state';
 import toJS from '../../../../HOCs/state/toJS';
 
 // Constants
